update: 로고 컴포넌트 메뉴로 합치기
