+++ src/components/layout/Menu/styles.ts
@@ -1,13 +1,60 @@
 import styled, { css } from 'styled-components';
 import darkLightIcon from 'images/icon-dark-light.png';
 import searchIcon from 'images/icon-search.png';
+import logoDark from 'images/icon-logo-dark.png';
+import logoLight from 'images/icon-logo-light.png';
+import { ColorModeState } from 'atoms/colorMode';
 
-interface IconProps {
+interface darkModeIconProps {
   searchButtonDisplay: string;
 }
+
+export const MainHeader = styled.section`
+  text-align: center;
+
+  @media (max-width: 600px) {
+    padding-top: 10px;
+  }
+`;
+export const Logo = styled.div`
+  justify-content: center;
+  display: flex;
+  align-items: center;
+`;
+
+export const Image = styled.img`
+  width: 230px;
+  height: 40px;
+`;
+
+interface IconProps {
+  colorMode: ColorModeState;
+}
+
+const getLogoIconSrc = (colorMode: ColorModeState) =>
+  colorMode === 'DARK' ? logoLight.src : logoDark.src;
+
+const Icon = styled.i`
+  display: block;
+  width: 230px;
+  height: 40px;
+  background-repeat: no-repeat;
+  background-size: contain;
+  background-position: center, center;
+`;
+
+export const Img = styled(Icon)<IconProps>`
+  background-image: url('${({ colorMode }) => getLogoIconSrc(colorMode)}');
+  width: 230px;
+  height: 40px;
+  @media (max-width: 1023px) {
+    width: 172px;
+    height: 30px;
+  }
+`;
+
 export const Main = styled.section`
   text-align: center;
-  margin-top: 30px;
   padding-bottom: 40px;
   @media (max-width: 1023px) {
     margin-top: 10px;
@@ -44,7 +91,7 @@ export const Button = styled.button`
   }
 `;
 
-const Icon = styled.i<IconProps>`
+const darkModeIcon = styled.i<darkModeIconProps>`
   display: ${({ searchButtonDisplay }) => searchButtonDisplay};
   width: 20px;
   height: 20px;
@@ -53,10 +100,10 @@ const Icon = styled.i<IconProps>`
   background-position: center, center;
 `;
 
-export const SearchIcon = styled(Icon)<IconProps>`
+export const SearchIcon = styled(darkModeIcon)<darkModeIconProps>`
   background-image: url('${searchIcon.src}');
 `;
 
-export const DarkLightIcon = styled(Icon)`
+export const DarkLightIcon = styled(darkModeIcon)`
   background-image: url('${darkLightIcon.src}');
 `;
